+++ Hyperiso/Hyperiso/pyhyperiso/Core.py
@@ -2,9 +2,10 @@
 
 from pyhyperiso.core.Core.HyperisoMaster import HyperisoMaster
 from pyhyperiso.core.Core.HyperisoConfig import HyperisoConfig, ExternalFlag
-from pyhyperiso.core.Core.ParamaterProvider import ParameterProvider
+from pyhyperiso.core.Core.ParameterProvider import ParameterProvider
 from pyhyperiso.core.Core.ParameterSetter import ParameterSetter
 from pyhyperiso.core.Core.BlockProvider import BlockLogger
+from pyhyperiso.core.Core.DatabaseWriter import DatabaseWriter
 from pyhyperiso.core.Core.DependencyPruner import DependencyPruner
 from pyhyperiso.core.Core.DependantBlockInfoProvider import DependantBlockInfoProvider
 from pyhyperiso.core.Core.CorrelationProvider import CorrelationProvider